Started BaseMapManager

+++ src/ch/epfl/javelo/gui/MapViewParameters.java
@@ -4,7 +4,6 @@ import ch.epfl.javelo.projection.PointWebMercator;
 import javafx.geometry.Point2D;
 
 public record MapViewParameters(int zoomlevel, double originXcoord, double originYcoord) {
-
     public Point2D topLeft(){
         return new Point2D(originXcoord, originYcoord);
     }
@@ -19,11 +18,11 @@ public record MapViewParameters(int zoomlevel, double originXcoord, double origi
         return PointWebMercator.of(zoomlevel, x, y);
     }
 
-    public int viewX(PointWebMercator point){
-        return (int)(point.xAtZoomLevel(zoomlevel) - originXcoord);
+    public double viewX(PointWebMercator point){
+        return (point.xAtZoomLevel(zoomlevel) - originXcoord);
     }
 
-    public int viewY(PointWebMercator point){
-        return (int)(point.yAtZoomLevel(zoomlevel) - originYcoord);
+    public double viewY(PointWebMercator point){
+        return (point.yAtZoomLevel(zoomlevel) - originYcoord);
     }
 }
